tests: test implementation of tagged categorized cache
In this proof-of-concept commit, we have a database of books,
but only one category (Sci-Fi) is monitored and cached.

This is achieved by twofold keying:

* The cache itself is keyed by `category`
* a `dep.Tag()` is installed with name=category
* The `add_book()` method knows that categories are being watched,
  and invalidates them accordingly.

This approach works, but it increases hidden links within the code.
That is, `list_books()` and `add_book()` have to be aware of one another
and work in perfect tandem for this caching to work.

Supporting the integrity of such links manually is absolutely impossible
in real world, therefore, we have to find some declarative approach.