Scope: invalidate all caches in case of a failure in the extractor function
What shall we do if the extractor function has an exception?
In development mode, we of course report the error because we want
a developer to fix in.
In production mode, however, we understand that cache is an optional
facility and can be safely invalidated. It's better than a failure.