redis: fixed racing conditions by properly using Redis transactions
Fixed both methods, put() and invalidate().
Now it wraps every significant operation into `redis.pipeline()`,
which, by default starts a transaction.

In the pipeline, I use sub-`pipeline(transaction=False)` to fetch
many values as once. This is not a transaction but a bulk-load.

In the end, everything's committed atomically, using `MULTI + ... + EXEC`.

The whole thing has to be wrapped into a `while True` loop to retry the
transaction until it works. It's too scary to write loops like this,
so I used a `for _ in range(10)` instead.